Rename translate helper to i18n and localize remaining UI literals
