+++ src/ui/ResultTabs/ExcelCopyButton.jsx
@@ -1,5 +1,5 @@
 import { useState } from 'react'
-import { t } from '../../localization/translate.js'
+import { t } from '../../localization/i18n.js'
 import { Button, Tooltip } from '@mui/material'
 import { ContentCopyOutlined, CheckOutlined, ErrorOutlineOutlined } from '@mui/icons-material'
 import { ExcelPresenter } from '../../presentation/ExcelPresenter.js'
